Send file using fexsend

Documentation available at https://fex.belwue.de/fstools/fexsend.html

fex(..., zip_file = sprintf("%s_archive.zip", Sys.Date()), internal = TRUE)

Arguments

...
One or several paths to files. A zip archive is created when several paths are provided,

zip_file
Name of the archive created when several paths are provided. Default "20XX-XX-XX_archive.zip".

internal
If internal fexsend binary should be used. Default TRUE.

Value

The URL of the upload file.

Examples

if (interactive()) {
  writeLines("file1", file.path(tempdir(), "file1.txt"))
  writeLines("file2", file.path(tempdir(), "file2.txt"))

  fex(file.path(tempdir(), "file1.txt")) # File sent as-is
  fex(file.path(tempdir(), c("file1.txt", "file2.txt"))) # Files sent as a ZIP archive
}